% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxPlotMarkerMetaClust.R
\name{BoxPlotMarkerMetaClust}
\alias{BoxPlotMarkerMetaClust}
\title{Box plot of one marker MFI, for the different meta-cluster with statical analysis, output in a PDF file.}
\usage{
BoxPlotMarkerMetaClust(
  TreeMetaCl,
  Title,
  treatmentTable,
  ControlTreatment,
  BottomMargin,
  Marker,
  Robust = TRUE,
  ClustHeat = TRUE,
  ExportData = FALSE
)
}
\arguments{
\item{TreeMetaCl}{FlowSOM tree with meta-clusters, constructed by buildFSOMTree}

\item{Title}{prefix of the pdf file name}

\item{treatmentTable}{data frame containing a column 'files' and a column 'Treatment' (case sensitive, avoid special character, especially "-")}

\item{BottomMargin}{size of the boxplots bottom margins, in order to be readable}

\item{Marker}{marker name}

\item{Robust}{if TRUE, robust statistics is applied (dunn), otherwise linear-model/Tukey}

\item{ClustHeat}{if TRUE, heatmaps are clustered}

\item{ExportData}{if TRUE, export MFI,Median of MFI, pairwise comparison p-values, controle vs treatment p-values to csv files}

\item{Controltreatment}{name of the experimental condition that you consider as your control treatment used for statistics}
}
\value{
list containing MFI,Median of MFI, pairwise comparison p-values, controle vs treatment p-values
}
\description{
BoxPlotMarkerMetaClust is a function that compares, inside each metacluster, the median fluorescence intensity of one given marker for each experimental condition.
}
\examples{
BoxPlotMarkerMetaClust(TreeMetaCl=CytoTree, Title="Experiment1",treatmentTable=Experiment1table.csv,ControlTreatment="control",BottomMargin=3,Marker="PD-1",Robust=TRUE, ClustHeat=TRUE)
}
