% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxPlotMetaClust.R
\name{BoxPlotMetaClust}
\alias{BoxPlotMetaClust}
\title{Box plot of maker sizes, for the different meta-cluster with statical analysis, output in a PDF file.}
\usage{
BoxPlotMetaClust(
  TreeMetaCl,
  Title,
  treatmentTable,
  ControlTreatment,
  BottomMargin,
  yLab,
  Norm = FALSE,
  Robust = TRUE,
  ClustHeat = TRUE
)
}
\arguments{
\item{TreeMetaCl}{FlowSOM tree with meta-clusters, constructed by buildFSOMTree}

\item{Title}{prefix of the pdf file name}

\item{treatmentTable}{data frame containing a column 'files' and a column 'Treatment' (case sensitive)}

\item{BottomMargin}{size of the boxplots bottom margins, in order to be readable}

\item{Norm}{if true, sizes are normlized according to column 'NormalizationFactor' of the treatment table. Otherwise, percentage is used}

\item{Robust}{if TRUE, robust statistics is applied (dunn), otherwise linear-model/Tukey}

\item{ClustHeat}{if TRUE, heatmaps are clustered}

\item{Controltreatment}{name of the control treatment used for statistics}

\item{ylab}{name of y-axis of boxplots (eg 'population size')}
}
\value{
list containing sizes, pairwise comparison p-values, controle vs treatment p-values
}
\description{
Box plot of maker sizes, for the different meta-cluster with statical analysis, output in a PDF file.
}
