% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFSOMTree.R
\name{buildFSOMTree}
\alias{buildFSOMTree}
\title{build FSOM tree with the metacluster and plot the tree}
\usage{
buildFSOMTree(fSOMDloaded, prettyNames, clustDim, metaClNb, fSOMSeed)
}
\arguments{
\item{prettyNames}{list of pretty names needed for building the tree}

\item{clustDim}{dimension of the 2D cluster grid}

\item{fSOMSeed}{seed of the random generator used for building the tree}

\item{fSOMloaded}{FlowSOM data dowloaded with DowLoadCytoData}

\item{metaClNg}{number of meta-clusters}
}
\value{
FlowSOM tree with meta-clusters
}
\description{
build FSOM tree with the metacluster and plot the tree
}
