% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFromMetaClust.R
\name{DataFromMetaClust}
\alias{DataFromMetaClust}
\title{Construct a sub-dataset from dataset and a list of metaclusters}
\usage{
DataFromMetaClust(FSOMData, TreeMetaCl, MetaClusters)
}
\arguments{
\item{FSOMData}{FlowSOM datobject, created from FlowSOM::ReadInput, DownloadCytoData ($fSOMData element), PolygonGatingRawData ($fSOMData element) or PolygonGatingGatedData ($fSOMData element)}

\item{TreeMetaCl}{FlowSOM tree with meta-clusters, constructed within buildFSOMTree (from the data of FSOMData)}

\item{MetaClusters}{vector of metaclusters numbers or names}
}
\value{
FlowSOM data object
}
\description{
Construct a sub-dataset from dataset and a list of metaclusters
}
