% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistMarkerMetaCl.R
\name{HistMarkerMetaCl}
\alias{HistMarkerMetaCl}
\title{Hitogram representation of a marker in a given meta-cluster, compared to the full data}
\usage{
HistMarkerMetaCl(
  TreeMetaCl,
  Marker,
  MetaCl,
  breaks = 100,
  xlim = NULL,
  freq = TRUE
)
}
\arguments{
\item{TreeMetaCl}{Object generated by the function BuildFSOMTree}

\item{Marker}{Marker name}

\item{MetaCl}{a meta-cluster name or number}

\item{breaks}{parameter to pass to R hist function}

\item{xlim}{parameter to pass to R hist function}

\item{freq}{paramter to pass to R hist function}
}
\value{
a double histogram figure, with meta-cluster in red
}
\description{
Hitogram representation of a marker in a given meta-cluster, compared to the full data
}
