% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractivePolyGate.R
\name{InteractivePolyGate}
\alias{InteractivePolyGate}
\title{interactive gating from data. Mouse click to build polygon, right click to exit.}
\usage{
InteractivePolyGate(
  RawData,
  marker1,
  marker2,
  fcsFiles,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{RawData}{data from FlowSOM::ReadInput}

\item{marker1}{x-axis}

\item{marker2}{y-axis}

\item{fcsFiles}{list of file names or list of file indices to be used}

\item{xlim}{limits of x-axis, if NULL, limits set automatically}

\item{ylim}{lmiits of y-axis, if NULL, limits set automatically}
}
\value{
a polygon with the marker names
}
\description{
interactive gating from data. Mouse click to build polygon, right click to exit.
}
