% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot2DScatterPlot.R
\name{Plot2DScatterPlot}
\alias{Plot2DScatterPlot}
\title{Plot2DScatterPlot}
\usage{
Plot2DScatterPlot(TreeMetaCl, Title, Marker1, Marker2, mincnt, style)
}
\arguments{
\item{TreeMetaCl}{object generated by buildFSOMTree function}

\item{Title}{title of the 2D scatter plot}

\item{Marker1}{marker to plot on the abscissa ("x")}

\item{Marker2}{marker to plot on the ordinate ("y")}

\item{style}{appareance of the polygons for the plot, that can be "colorscape","nestle.lattice", "nestle.centroids"...}

\item{minct}{minimun count of cells to display as a polygon on the plot}
}
\value{
PDF file of the chosen 2D scatter plot, that is saved in the working directory
}
\description{
Plot2DScatterPlot displays a 2D scatter plot of specified markers using the hexbin package.
}
\examples{
Plot2DScatterPlot(CytoTree,"Experiment 1","FSC-A","SSC-A",100,"colorscale")
}
