% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLabelsRm.R
\name{PlotLabelsRm}
\alias{PlotLabelsRm}
\title{Modified PlotLabels function of FlowSOM, with removal of smallest clusters}
\usage{
PlotLabelsRm(fSOMObject, metaClustFactors, mainTitle, nbRm, equalSize = F)
}
\arguments{
\item{fSOMObject}{FlowSOM tree}

\item{metaClustFactors}{meta-clusters (numbers or names)}

\item{nbRm}{number of smallest cluster to remove}

\item{equalSize}{true if clusters are represented with identical sizes}
}
\description{
Modified PlotLabels function of FlowSOM, with removal of smallest clusters
}
