% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMarkerMSTRm.R
\name{PlotMarkerMSTRm}
\alias{PlotMarkerMSTRm}
\title{Modified PlotMarkerMST function of FlowSOM, with removal of smallest clusters}
\usage{
PlotMarkerMSTRm(fSOMObject, markerName, mainTitle, nbRm, globalMinMax = c())
}
\arguments{
\item{fSOMObject}{FlowSOM tree}

\item{markerName}{name of marker}

\item{mainTitle}{title}

\item{nbRm}{number of smallest cluster to remove}

\item{globalMinMax}{min and max value of markers, used for color scale}
}
\description{
Modified PlotMarkerMST function of FlowSOM, with removal of smallest clusters
}
