% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotStarsMSTRm.R
\name{PlotStarsMSTRm}
\alias{PlotStarsMSTRm}
\title{Modified PlotStarsMST function of FlowSOM, with removal of smallest clusters}
\usage{
PlotStarsMSTRm(fSOMObject, metaClustFactors, mainTitle, nbRm = 0)
}
\arguments{
\item{fSOMObject}{FlowSOM tree}

\item{metaClustFactors}{meta-clusters (numbers or names)}

\item{mainTitle}{title}

\item{nbRm}{number of smallest cluster to remove}
}
\description{
Modified PlotStarsMST function of FlowSOM, with removal of smallest clusters
}
