% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTreeSet.R
\name{plotTreeSet}
\alias{plotTreeSet}
\title{Plot a set of trees, in a pdf file, for metaclusters and markers}
\usage{
plotTreeSet(
  TreeMetacl,
  markers,
  Title,
  rmClNb = 0,
  treatmentTable,
  globalScale = T
)
}
\arguments{
\item{TreeMetacl}{FlowSOM tree with meta-clusters, constructed by buildFSOMTree}

\item{markers}{vector of markers}

\item{Title}{title}

\item{rmClNb}{number of smallest clusters to remove}

\item{treatmentTable}{data frame containing a column 'files' and a column 'Treatment' (case sensitive)}

\item{globalScale}{TRUE if color scale for markers is the same for each treatments}
}
\description{
Plot a set of trees, in a pdf file, for metaclusters and markers
}
