% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolygonGatingRawData.R
\name{PolygonGatingRawData}
\alias{PolygonGatingRawData}
\title{Apply a list of gates made of polygon (from function Interactive2DGating) to RawData}
\usage{
PolygonGatingRawData(RawData, Polygons, gatingName = "")
}
\arguments{
\item{RawData}{data from FlowSOM::ReadInput}

\item{Polygons}{list of polygons contructed within Interactive2DGating}

\item{gatingName}{Name of gating}
}
\value{
FlowSOM data object that includes the polygons gating and the name of the population of interest.
}
\description{
Apply a list of gates made of polygon (from function Interactive2DGating) to RawData
}
