% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterPlotMatrix.R
\name{ScatterPlotMatrix}
\alias{ScatterPlotMatrix}
\title{ScatterPlotMatrix}
\usage{
ScatterPlotMatrix(TreeMetaCl, Title)
}
\arguments{
\item{TreeMetaCl}{object generated by buildFSOMTree function}

\item{Title}{title of the matrix}
}
\value{
PDF file of the matrix plot saved in the working directory
}
\description{
ScatterPlotMatrix creates a matrix of scatter plot through the "pairs" function displaying the raw data pool of all fcs files.
}
\examples{
ScatterPlotMatrix(CytoTree,"Experiment 1")
}
