% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectFilesGatedData.R
\name{SelectFilesGatedData}
\alias{SelectFilesGatedData}
\title{Select a new data from a list of files}
\usage{
SelectFilesGatedData(GatedData, Files)
}
\arguments{
\item{GatedData}{data from CytoSOM::DownloadCytoData or CytoSOM::PolygonGatingGatedData}

\item{Files}{list of file names (with no directory name)}
}
\value{
FlowSOM data object.
}
\description{
Select a new data from a list of files
}
