% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectFilesRawData.R
\name{SelectFilesRawData}
\alias{SelectFilesRawData}
\title{Select a new data from a list of files}
\usage{
SelectFilesRawData(RawData, Files)
}
\arguments{
\item{RawData}{data from FlowSOM::ReadInput}

\item{Files}{list of file names (with no directory name)}
}
\value{
FlowSOM data object.
}
\description{
Select a new data from a list of files
}
