% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMetaManualRenaming.R
\name{TreeMetaManualRenaming}
\alias{TreeMetaManualRenaming}
\title{Selective naming of metaclustesters according to specific markers expression}
\usage{
TreeMetaManualRenaming(TreeMetaCl, NamesDF)
}
\arguments{
\item{TreeMetaCl}{Object generated by the function BuildFSOMTree}

\item{NameDF}{A data frame with two column: "oldName" and "newName". No identical names in column newName is allowed. Avoid factors.}
}
\value{
BuildFSOMTree object with the metaclusters renamed
}
\description{
TreeMetaRenaming is a function that allows the automatic renaming of the metaclusters according to their expression for specific markers. If this function is applied, the metaclusters will be renamed for the following commands of the script.
}
