% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxPlotMarkerMetaClust.R
\name{BoxPlotMarkerMetaClust}
\alias{BoxPlotMarkerMetaClust}
\title{Box plot of maker MFI, for the different meta-cluster with statical analysis, output in a PDF file.}
\usage{
BoxPlotMarkerMetaClust(
  TreeMetaCl,
  Title,
  treatmentTable,
  ControlTreatment,
  BottomMargin,
  Marker,
  Robust = TRUE,
  ClustHeat = TRUE
)
}
\arguments{
\item{TreeMetaCl}{FlowSOM tree with meta-clusters, constructed by buildFSOMTree}

\item{Title}{prefix of the pdf file name}

\item{treatmentTable}{data frame containing a column 'files' and a column 'Treatment' (case sensitive)}

\item{BottomMargin}{size of the boxplots bottom margins, in order to be readable}

\item{Marker}{marker name}

\item{Robust}{if TRUE, robust statistics is applied (dunn), otherwise linear-model/Tukey}

\item{ClustHeat}{if TRUE, heatmaps are clustered}

\item{Controltreatment}{name of the control treatment used for statistics}
}
\value{
list containing MFI, pairwise comparison p-values, controle vs treatment p-values
}
\description{
Box plot of maker MFI, for the different meta-cluster with statical analysis, output in a PDF file.
}
