% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxPlotMetaClust.R
\name{BoxPlotMetaClust}
\alias{BoxPlotMetaClust}
\title{Statistical comparison of metacluster population between experimental condition}
\usage{
BoxPlotMetaClust(
  TreeMetaCl,
  Title,
  treatmentTable,
  ControlTreatment,
  BottomMargin,
  yLab,
  Norm = FALSE,
  Robust = TRUE,
  ClustHeat = TRUE,
  ExportData = FALSE
)
}
\arguments{
\item{TreeMetaCl}{FlowSOM tree with meta-clusters, constructed by buildFSOMTree}

\item{Title}{prefix of the pdf file name}

\item{treatmentTable}{data frame containing a column 'files',a column 'Treatment' (case sensitive, avoid special character, especially "-") and a column "NormalizationFactor" for the optional normalization of the data}

\item{BottomMargin}{size of the boxplots bottom margins, in order to be readable}

\item{Norm}{if true, sizes are normlized according to column 'NormalizationFactor' of the treatment table. Otherwise, percentage is used}

\item{Robust}{if TRUE, robust statistics is applied (dunn), otherwise linear-model/Tukey}

\item{ClustHeat}{if TRUE, heatmaps are clustered}

\item{ExportData}{if TRUE, export sizes,median of sizes, pairwise comparison p-values, controle vs treatment p-values to csv files}

\item{Controltreatment}{name of the control treatment used for statistics}

\item{ylab}{name of y-axis of boxplots (eg 'population size')}
}
\value{
list containing sizes,median of sizes, pairwise comparison p-values, controle vs treatment p-values
}
\description{
BoxPlotMetaClust is a function that allow the statistical analysis, either in percentage or with a normalization, of the impact of experimental condition on each metacluster population.
Box plot of maker sizes, for the different meta-cluster with statical analysis, output in a PDF file.
}
\examples{
BoxplotMetaClust(TreeMetaCl=CytoTree, Title="Experiment1",treatmentTable="ExperimentTable1.csv",ControlTreatment="control",BottomMargin=3,yLab="CD45+",Norm=FALSE,Robust=TRUE,ClustHeat=TRUE)
}
