% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFSOMTree.R
\name{buildFSOMTree}
\alias{buildFSOMTree}
\title{build FSOM tree with the metacluster and plot the tree}
\usage{
buildFSOMTree(fSOMDloaded, prettyNames, clustDim, metaClNb, fSOMSeed)
}
\arguments{
\item{prettyNames}{List of pretty names needed for building the tree}

\item{clustDim}{Dimension of the 2D cluster grid}

\item{fSOMSeed}{Seed of the random generator used for building the tree}

\item{fSOMloaded}{FlowSOM data dowloaded with DowLoadCytoData}

\item{metaClNg}{Number of meta-clusters}
}
\value{
A FlowSOM tree
}
\description{
build FSOM tree with the metacluster and plot the tree
}
