% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFromMetaClust.R
\name{DataFromMetaClust}
\alias{DataFromMetaClust}
\title{Construct a sub-dataset from dataset and a list of metaclusters}
\usage{
DataFromMetaClust(FSOMData, TreeMetaCl, MetaClusters)
}
\arguments{
\item{FSOMData}{FlowSOM datobject, created from DownLoadCytoData}

\item{TreeMetaCl}{FlowSOM tree with meta-clusters, constructed within buildFSOMTree}

\item{MetaClusters}{vector of metaclusters numbers or names}
}
\value{
FlowSOM data object
}
\description{
Construct a sub-dataset from dataset and a list of metaclusters
}
