% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownLoadCytoData.R
\name{DownLoadCytoData}
\alias{DownLoadCytoData}
\title{Download data, given fcs and FlowJo workspace}
\usage{
DownLoadCytoData(
  dirFCS = "",
  gatingName,
  fcsPattern = "Tube",
  compensate = FALSE
)
}
\arguments{
\item{dirFCS}{sub-directory of working directory, where fcs files are}

\item{gatingName}{name fo FlowJo gating to be applied}

\item{fcsPattern}{pattern of all fcs file names}

\item{compensate}{Does the data need to be compensated?}
}
\value{
A FlowSOM data object
}
\description{
Download data, given fcs and FlowJo workspace
}
