% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindMetaClustNames.R
\name{FindMetaClustNames}
\alias{FindMetaClustNames}
\title{Find the names of metaclusters from a sub-name}
\usage{
FindMetaClustNames(subName, TreeMetaCl)
}
\arguments{
\item{subName}{sub-part of a meta-cluster name, for meta-cluster number preceeding its name, use "^number_"}

\item{TreeMetaCl}{FlowSOM tree with meta-clusters, constructed within buildFSOMTree (from the data of FSOMData)}
}
\value{
a vector of names
}
\description{
Find the names of metaclusters from a sub-name
}
