% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatMapTree.R
\name{HeatMapTree}
\alias{HeatMapTree}
\title{Heatmap representation of markers median values in meta-clusters.}
\usage{
HeatMapTree(TreeMetaCl, Markers, clustering = T)
}
\arguments{
\item{TreeMetaCl}{Object generated by the function BuildFSOMTree}

\item{Markers}{vector of Markers}

\item{clustering}{if true, a hierarichal clustring is applied}
}
\value{
a heatmap figure
}
\description{
Heatmap representation of markers median values in meta-clusters.
}
