% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaClusterNaming.R
\name{MetaClusterNaming}
\alias{MetaClusterNaming}
\title{Naming of meta-cluster from median and quartiles}
\usage{
MetaClusterNaming(TreeMetaCl, Markers)
}
\arguments{
\item{TreeMetaCl}{FlowSOM tree, contructed for buildFSOMTree}

\item{Markers}{vector of markers used for naming}
}
\value{
data frame with cluster number, non-robust name, robust name and short robust names
}
\description{
Naming of meta-cluster from median and quartiles
}
