% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMarkerMSTCondRm.R
\name{PlotMarkerMSTCondRm}
\alias{PlotMarkerMSTCondRm}
\title{Modified PlotMarkerMST function of FlowSOM, with removal of smallest clusters, over a subset of data}
\usage{
PlotMarkerMSTCondRm(
  fSOMObject,
  markerName,
  condIndex,
  mainTitle,
  nbRm,
  globalMinMax = c()
)
}
\arguments{
\item{fSOMObject}{FlowSOM tree}

\item{markerName}{name of marker}

\item{condIndex}{index of data events}

\item{mainTitle}{title}

\item{nbRm}{number of smallest cluster to remove}

\item{globalMinMax}{min and max value of markers, used for color scale}
}
\description{
Modified PlotMarkerMST function of FlowSOM, with removal of smallest clusters, over a subset of data
}
