% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotStarsMSTCondRm.R
\name{PlotStarsMSTCondRm}
\alias{PlotStarsMSTCondRm}
\title{Modified PlotStarsMST function of FlowSOM, with removal of smallest clusters, over a subset of data}
\usage{
PlotStarsMSTCondRm(
  fSOMObject,
  metaClustFactors,
  condIndex,
  mainTitle,
  nbRm,
  smallTree = F
)
}
\arguments{
\item{fSOMObject}{FlowSOM tree}

\item{metaClustFactors}{meta-clusters (numbers or names)}

\item{condIndex}{index of data events}

\item{mainTitle}{title}

\item{nbRm}{number of smallest cluster to remove}

\item{smallTree}{true if tree is small and meta-cluster legend is big}
}
\description{
Modified PlotStarsMST function of FlowSOM, with removal of smallest clusters, over a subset of data
}
