% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMetaRenaming.R
\name{TreeMetaRenaming}
\alias{TreeMetaRenaming}
\title{Automatic naming of metaclustesters according to specific markers expression}
\usage{
TreeMetaRenaming(TreeMetaCl, Markers, Levels, CutOff = NULL, Short = FALSE)
}
\arguments{
\item{TreeMetaCl}{Object generated by the function BuildFSOMTree}

\item{Markers}{vector of markers used for naming}

\item{Levels}{number of level for naming: 2 -> -/+ (1/2-4 quartiles), 3-> -/low/high (1/2-3/4 quartiles), 4-> -/low/int/high (1/2/3/4 quartiles)}

\item{CutOff}{matrix for optional cut off values for naming: one line for each markers (in the same order), each column for separating levels. If NULL, quartiles are used}

\item{Short}{is true, low is not represented with Levels=3, int is not represented if Levels = 4}
}
\value{
BuildFSOMTree object with the metaclusters renamed with the choosen name type, keeping the metacluster numbers ($metaClNumber)
}
\description{
A name based of markers level is added to the metcluster names (of number, in that case, the numbers are kept in the object as $metaClNumber)
}
\examples{
CytoTreeNewNames <- TreeMetaRenaming(TreeMetaCl=CytoTree,Markers=c("CD4","CD8","FoxP3"),Levels=4)
}
