% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMetaRenaming.R
\name{TreeMetaRenaming}
\alias{TreeMetaRenaming}
\title{Selective naming of metaclustesters according to specific markers expression}
\usage{
TreeMetaRenaming(TreeMetaCl, Markers, NameType)
}
\arguments{
\item{TreeMetaCl}{Object generated by the function BuildFSOMTree}

\item{Markers}{Markers choosen for the naming}

\item{NameType}{Type of name applied}
}
\value{
BuildFSOMTree object with the metaclusters renamed with the choosen name type
}
\description{
Selective naming of metaclustesters according to specific markers expression
}
